DDF-6111 Changes the CSRF filter from being a servlet filter to a security filter to prevent future race condition
