+++ distribution/ddf-common/src/main/resources/security/default.policy
@@ -307,6 +307,10 @@ grant codeBase "file:/admin-core-jolokia/javax.servlet-api/org.apache.shiro.core
     permission java.security.SecurityPermission "insertProvider";
 }
 
+grant codeBase "file:/security-filter-csrf" {
+    permission java.io.FilePermission "${ddf.home.perm}etc${/}custom.system.properties", "read";
+}
+
 grant codeBase "file:/security-core-api/security-core-services/security-handler-saml" {
     permission java.io.FilePermission "${ddf.home.perm}etc${/}ws-security${/}server${/}-", "read";
     permission java.io.FilePermission "${ddf.home.perm}etc${/}ws-security${/}issuer${/}-", "read";
