+++ features/security/pom.xml
@@ -190,6 +190,11 @@
         </dependency>
         <dependency>
             <groupId>ddf.security.filter</groupId>
+            <artifactId>security-filter-csrf</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>ddf.security.filter</groupId>
             <artifactId>security-filter-authorization</artifactId>
             <version>${project.version}</version>
         </dependency>
