+++ features/security/src/main/feature/feature.xml
@@ -410,6 +410,11 @@ org.ops4j.pax.web.default.realmname=DDF
         <bundle>mvn:ddf.security.filter/security-filter-login/${project.version}</bundle>
     </feature>
 
+    <feature name="security-filter-csrf" version="${project.version}"
+             description="CSRF Filter for web applications.">
+        <bundle>mvn:ddf.security.filter/security-filter-csrf/${project.version}</bundle>
+    </feature>
+
     <feature name="security-filter-authz" version="${project.version}"
              description="Authz Filter for web applications.">
         <feature>security-policy-context</feature>
@@ -581,6 +586,7 @@ org.ops4j.pax.web.default.realmname=DDF
         <feature>security-core-misc</feature>
         <feature>security-filter-web-sso</feature>
         <feature>security-filter-login</feature>
+        <feature>security-filter-csrf</feature>
         <feature>security-handler-basic</feature>
         <feature>security-oidc-realm</feature>
         <feature>security-handler-oidc</feature>
