+++ platform/security/filter/pom.xml
@@ -32,6 +32,7 @@
     <modules>
         <module>security-filter-web-sso</module>
         <module>security-filter-login</module>
+        <module>security-filter-csrf</module>
         <module>security-filter-authorization</module>
     </modules>
 </project>
