+++ platform/security/filter/security-filter-csrf/src/main/java/org/codice/ddf/security/filter/csrf/CsrfFilter.java
@@ -11,7 +11,7 @@
  * License is distributed along with this program and can be found at
  * <http://www.gnu.org/licenses/lgpl.html>.
  */
-package org.codice.ddf.pax.web.jetty;
+package org.codice.ddf.security.filter.csrf;
 
 import ddf.security.audit.SecurityLogger;
 import java.io.IOException;
@@ -24,10 +24,6 @@ import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import java.util.regex.Pattern;
-import javax.servlet.Filter;
-import javax.servlet.FilterChain;
-import javax.servlet.FilterConfig;
-import javax.servlet.ServletException;
 import javax.servlet.ServletRequest;
 import javax.servlet.ServletResponse;
 import javax.servlet.http.HttpServletRequest;
@@ -37,12 +33,16 @@ import javax.ws.rs.core.MultivaluedHashMap;
 import javax.ws.rs.core.MultivaluedMap;
 import org.apache.commons.lang.StringUtils;
 import org.codice.ddf.configuration.SystemBaseUrl;
+import org.codice.ddf.platform.filter.AuthenticationException;
+import org.codice.ddf.platform.filter.AuthenticationFailureException;
+import org.codice.ddf.platform.filter.FilterChain;
+import org.codice.ddf.platform.filter.SecurityFilter;
 import org.eclipse.jetty.http.HttpMethod;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 /** Contains multiple checks to prevent cross-site requests for protected context paths. */
-public class CsrfFilter implements Filter {
+public class CsrfFilter implements SecurityFilter {
 
   public static final String CSRF_HEADER = "X-Requested-With";
   public static final String ORIGIN_HEADER = "Origin";
@@ -78,8 +78,10 @@ public class CsrfFilter implements Filter {
 
   private SecurityLogger securityLogger;
 
-  public CsrfFilter() {
+  public CsrfFilter(SecurityLogger securityLogger) {
     super();
+    this.securityLogger = securityLogger;
+
     trustedAuthorities = new ArrayList<>();
     protectedContexts = new ArrayList<>();
 
@@ -89,7 +91,7 @@ public class CsrfFilter implements Filter {
   }
 
   @Override
-  public void init(FilterConfig filterConfig) throws ServletException {
+  public void init() {
 
     shouldFilter = csrfEnabled();
 
@@ -119,7 +121,7 @@ public class CsrfFilter implements Filter {
     trustedAuthorities.add(externalHostname + ":" + externalHttpPort);
     trustedAuthorities.add(externalHostname + ":" + externalHttpsPort);
 
-    // administator trusted authorities
+    // administrator trusted authorities
     List<String> administratorTrustedAuthorities = getAdministratorTrustedAuthorities();
     trustedAuthorities.addAll(administratorTrustedAuthorities);
 
@@ -156,11 +158,11 @@ public class CsrfFilter implements Filter {
    * @param response response stream for returning the response
    * @param chain chain of filters to be invoked following this filter
    * @throws IOException
-   * @throws ServletException
+   * @throws AuthenticationException
    */
   @Override
   public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain)
-      throws IOException, ServletException {
+      throws IOException, AuthenticationException {
 
     if (shouldFilter) {
       LOGGER.debug("Performing doFilter() on CsrfFilter");
@@ -174,14 +176,14 @@ public class CsrfFilter implements Filter {
       if (protectedContexts.stream().anyMatch(targetContextPath::startsWith)
           && doBrowserProtectionFilter(
               httpRequest, httpResponse, targetContextPath, requestMethod)) {
-        return;
+        throw new AuthenticationFailureException();
       }
 
       // Execute CSRF check if user is accessing /services
       if (targetContextPath.startsWith(SERVICE_CONTEXT)
           && doSystemProtectionFilter(
               httpRequest, httpResponse, targetContextPath, requestMethod, userAgentHeader)) {
-        return;
+        throw new AuthenticationFailureException();
       }
     }
 
@@ -390,9 +392,5 @@ public class CsrfFilter implements Filter {
   @Override
   public void destroy() {
     LOGGER.debug("Destroying CSRF filter.");
-  }
-
-  public void setSecurityLogger(SecurityLogger securityLogger) {
-    this.securityLogger = securityLogger;
   }
 }
